Rename worker-routine.ts to worker.ts
